+++ dvc/repo/get_url.py
@@ -1,19 +1,17 @@
 import os
 
+from dvc import output
+from dvc.exceptions import URLMissingError
+from dvc.fs import download, parse_external_url
+from dvc.utils import resolve_output
+
 
 def get_url(url, out=None, jobs=None):
-    from dvc import dependency, output
-    from dvc.utils import resolve_output
-
     out = resolve_output(url, out)
-
-    if os.path.exists(url):
-        url = os.path.abspath(url)
-
     out = os.path.abspath(out)
-
-    (dep,) = dependency.loads_from(None, [url])
     (out,) = output.loads_from(None, [out], use_cache=False)
-    if not dep.exists:
-        raise dep.DoesNotExistError(dep)
-    dep.download(out, jobs=jobs)
+
+    src_fs, src_path = parse_external_url(url)
+    if not src_fs.exists(src_path):
+        raise URLMissingError(url)
+    download(src_fs, src_path, out, jobs=jobs)
